Offline kNN was not trully offline. Begining to fix that.
Scikit-learn call to fit does not pre-calculate distances or does any offline
pre-processing. My neighbors class was changed to support it. Recommender is
being adpted but has not finished.

Besides, BMFrecommender transform has changed (linear algebra is not directly
applicable to a boolean matrix product). It is working for BMFrecommender, but
crashes when called from BMFRPrecommender. Check on that latter.