Fixes in rating prediction stratgy for neighborhood methods.
1)Afected recommenders using zero_mean during prediction.
Denominator was considering only ratings above zero, but when using
zero-mean, many rating deviations are negative and should be accounted
for.

2) Afected recommenders using zero-mean before similarity computation
Denominator must be the sum of the modules of the similarities.
If vectors are not all positive, cosine similarity may output a negative value,
which makes the old denominator value wrong.