Ensemble using only model, not full recomendations data
if when performing recomendations the ensemble object has to access
a SavedRecommendations object for each base RS, it will keep loading
enormous amount of data onto RAM (saved recomendations is basically
a large dense matrix), which is causing a huge recommendation
time overload. Changed to use only the model and do predicitions on
the fly.